% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motionChart.R
\name{auto_label_size}
\alias{auto_label_size}
\title{Choose a default label font size based on entity count}
\usage{
auto_label_size(n)
}
\arguments{
\item{n}{Integer number of entities.}
}
\value{
Integer font size in pixels.
}
\description{
Fewer entities give larger labels; crowded charts give smaller labels.
The user can always override interactively with the A- / A+ buttons.
}
