% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwd.R
\name{openwd}
\alias{openwd}
\title{Open Windows Directory}
\usage{
openwd(myDocuments = FALSE, parentDirectory = FALSE, drive = NULL)
}
\arguments{
\item{myDocuments}{Logical. If TRUE, opens the user's Documents folder. 
Default is FALSE.}

\item{parentDirectory}{Logical. If TRUE, opens the parent directory of the 
current working directory. Default is FALSE.}

\item{drive}{Character. Drive letter (e.g., "C", "D") to open that drive's
root directory. If specified, other parameters are ignored. Default is NULL.}
}
\value{
No return value, opens Windows Explorer window.
}
\description{
Open a Windows File Explorer window at specified locations.
}
\details{
Uses system shell commands to open Windows Explorer.
}
\examples{
\dontrun{
# Open current working directory
openwd()

# Open user's Documents folder
openwd(myDocuments = TRUE)

# Open parent directory
openwd(parentDirectory = TRUE)

# Open C: drive
openwd(drive = "C")
}

}
