% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gapminder_dslabs.R
\name{load_gapminder_dslabs}
\alias{load_gapminder_dslabs}
\title{Load the DSLabs Gapminder dataset}
\usage{
load_gapminder_dslabs()
}
\value{
A data frame with columns: country, year, infant_mortality, life_expectancy, fertility, population, gdp, continent, region, gdpPercap
}
\description{
Load the DSLabs Gapminder dataset
}
